# Optimize decoding