--- a/utils_test.go
+++ b/utils_test.go
@@ -36,18 +36,3 @@ func TestDeferredMaxLengthSingle(t *testing.T) {
 		t.Fatal("deferred: allowed more than the maximum allocation supported")
 	}
 }
-
-func TestDeferredMaxLengthRecursive(t *testing.T) {
-	var header bytes.Buffer
-	for i := 0; i < MaxLength+1; i++ {
-		if err := WriteMajorTypeHeader(&header, MajTag, 0); err != nil {
-			t.Fatal("failed to write header")
-		}
-	}
-
-	var deferred Deferred
-	err := deferred.UnmarshalCBOR(&header)
-	if err != maxLengthError {
-		t.Fatal("deferred: allowed more than the maximum number of elements")
-	}
-}