This optimizes:

1. readByte: It turns out that frequently casting an interface to an interface is really slow, but casting an interface to a concrete type is fast. 
2. Deferred: This avoids most of the unnecessary allocations and copies in `Deferred.UnmarshalCBOR`, and switches to an iterative (rather than recursive) implementation.
3. ScanForLinks: As with Deferred, this is now non-recursive.

Note: depth limits have been removed from Deferred as the implementation is no longer recursive. ScanForLinks never appeared to have depth limits.