@Stebalien 2020-08-05
Based on #33.